create_vector_db/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.3.0
pdfplumber==0.10.3
qdrant-client==1.7.3
pypdf==4.0.1
llama_index==0.9.43
llama-hub==0.0.79.post1
weaviate-client==4.4.2
chromadb~=0.4.22
pymilvus==2.3.6
litellm==1.22.5
numpy~=1.24.3
pinecone-client==3.1.0
openai~=1.10.0
langchain==0.0.164
python-dotenv~=0.21.0
datasets~=2.12.0